Textfield component styled component migration
