+++ src/index.tsx
@@ -36,7 +36,7 @@ export { default as Stepper } from 'stepper'
 export { default as SuccessModal } from 'successModal'
 export { default as Tabs, TabStatus } from 'tabs'
 export { default as Text, TextDisplayType, TextTagType } from 'text'
-export { default as TextField } from 'textField'
+export { default as TextField, inputTypes, inputModes } from 'textField'
 export { default as TimePicker } from 'timePicker'
 export { default as Title } from 'title'
 export { default as ToggleButton } from 'toggleButton'
